Create unipos-suggest-members Element
